reset consumers #212
